move dex trade interface to separate file
